refactor: project structure and appearance
