+++ src/Components/Main/MainRoutes.tsx
@@ -1,23 +1,20 @@
 import type { FC, ReactElement } from 'react';
 import { Suspense, lazy } from 'react';
 import { Navigate, Route, Routes } from 'react-router-dom';
-import { NewsPage } from './Pages/NewsPage/NewsPage';
-import { MusicsPage } from './Pages/MusicsPage/MusicsPage';
-import { SettingsPage } from './Pages/SettingsPage/SettingsPage';
 import { useAuthRedirect } from '../../hooks/useRedirect';
 import { ProfilePageContainer } from './Pages/ProfilePage/ProfilePageContainer';
 import { Preloader } from '../Common/Preloader/Preloader';
 import { NotFoundPage } from './Pages/NotFoundPage/NotFoundPage';
 import { RoutePath } from '../../utils/types/enums';
 
-const MessagesPageContainer = lazy(() => import('./Pages/MessagesPage/MessagesPageContainer'));
+const DialogsPageContainer = lazy(() => import('./Pages/DialogsPage/DialogsPageContainer'));
 const UsersPageContainer = lazy(() => import('./Pages/UsersPage/UsersPageContainer'));
 const LoginPageContainer = lazy(() => import('./Pages/LoginPage/LoginPageContainer'));
 const ChatPage = lazy(() => import('./Pages/ChatPage/ChatPage'));
 
 export const MainRoutes: FC = (): ReactElement => {
   const authProfilePage = useAuthRedirect(<ProfilePageContainer />);
-  const authMessagesPage = useAuthRedirect(<MessagesPageContainer />);
+  const authDialogsPage = useAuthRedirect(<DialogsPageContainer />);
   const authChatPage = useAuthRedirect(<ChatPage />);
 
   return (
@@ -27,11 +24,8 @@ export const MainRoutes: FC = (): ReactElement => {
         <Route path={RoutePath.main} element={<Navigate to={RoutePath.profile} />} />
         <Route path={RoutePath.profile} element={authProfilePage} />
         <Route path={`${RoutePath.profile}/:id?`} element={<ProfilePageContainer />} />
-        <Route path={`${RoutePath.messages}/*`} element={authMessagesPage} />
+        <Route path={RoutePath.dialogs} element={authDialogsPage} />
         <Route path={RoutePath.users} element={<UsersPageContainer />} />
-        <Route path={RoutePath.news} element={<NewsPage />} />
-        <Route path={RoutePath.musics} element={<MusicsPage />} />
-        <Route path={RoutePath.settings} element={<SettingsPage />} />
         <Route path={RoutePath.login} element={<LoginPageContainer />} />
         <Route path={RoutePath.chat} element={authChatPage} />
       </Routes>
