+++ src/Components/Main/Pages/DialogsPage/DialogsPage.module.scss
