+++ src/Components/Main/Pages/DialogsPage/Message/Message.module.scss
