+++ src/Components/Main/Pages/DialogsPage/Post/Post.module.scss
