+++ src/Components/Main/Pages/DialogsPage/PostForm/PostForm.module.scss
