+++ src/Components/Main/Pages/DialogsPage/Posts/Posts.module.scss
