+++ src/Components/Main/Pages/DialogsPage/Posts/Posts.tsx
@@ -9,10 +9,9 @@ import { scrollToBottom } from '../../../../../utils/helpers/componentHelpers';
 type PropsType = {
   posts: Array<PostType>;
   addPost: (text: string) => void;
-  isAuth: boolean;
 };
 
-export const Posts: FC<PropsType> = ({ posts, addPost, isAuth }): ReactElement | boolean => {
+export const Posts: FC<PropsType> = ({ posts, addPost }): ReactElement | boolean => {
   const postsElements = posts.map((post: PostType): ReactElement => <Post postText={post.postText} key={post.id} />);
 
   const anchorRef = useRef<HTMLDivElement>(null);
@@ -21,14 +20,12 @@ export const Posts: FC<PropsType> = ({ posts, addPost, isAuth }): ReactElement |
   }, [posts]);
 
   return (
-    isAuth && (
-      <div className={classes.postsBlock}>
-        <h3 className={classes.title}>Posts</h3>
-        <PostFormContainer addPost={addPost} />
-        <div className={classes.posts} ref={anchorRef}>
-          {postsElements}
-        </div>
+    <div className={classes.postsBlock}>
+      <h3 className={classes.title}>Posts</h3>
+      <PostFormContainer addPost={addPost} />
+      <div className={classes.posts} ref={anchorRef}>
+        {postsElements}
       </div>
-    )
+    </div>
   );
 };
