--- src/Components/Main/Pages/MessagesPage/Dialog/Dialog.module.scss
@@ -1,16 +1,0 @@
-.dialog {
-  color: $content;
-  margin: 0 $marginMin $marginMin $marginMin;
-  cursor: pointer;
-  font-size: 1.2rem;
-  &:hover {
-    color: $captions;
-  }
-}
-.active {
-  color: $active;
-  cursor: default;
-  &:hover {
-    color: $active;
-  }
-}
