--- src/Components/Main/Pages/MessagesPage/Dialogs/Dialogs.module.scss
@@ -1,12 +1,0 @@
-.dialogsBlock {
-  width: 50%;
-  text-align: center;
-  border-right: 1px solid $active;
-  .title {
-    @extend %title;
-  }
-  .dialogs {
-    @include flex(null, center);
-    flex-direction: column;
-  }
-}
