--- src/Components/Main/Pages/MusicsPage/MusicsPage.module.scss
@@ -1,3 +1,0 @@
-.musicsPageBlock {
-  text-align: center;
-}
