--- src/Components/Main/Pages/MusicsPage/MusicsPage.tsx
@@ -1,6 +1,0 @@
-import type { FC, ReactElement } from 'react';
-import classes from './MusicsPage.module.scss';
-
-export const MusicsPage: FC = (): ReactElement => {
-  return <div className={classes.musicsPageBlock}>Musics</div>;
-};
