--- src/Components/Main/Pages/NewsPage/NewsPage.module.scss
@@ -1,3 +1,0 @@
-.newsPageBlock {
-  text-align: center;
-}
