--- src/Components/Main/Pages/NewsPage/NewsPage.tsx
@@ -1,6 +1,0 @@
-import type { FC, ReactElement } from 'react';
-import classes from './NewsPage.module.scss';
-
-export const NewsPage: FC = (): ReactElement => {
-  return <div className={classes.newsPageBlock}>News</div>;
-};
