--- src/Components/Main/Pages/SettingsPage/SettingsPage.module.scss
@@ -1,3 +1,0 @@
-.settingsPageBlock {
-  text-align: center;
-}
