--- src/Components/Main/Pages/SettingsPage/SettingsPage.tsx
@@ -1,6 +1,0 @@
-import type { FC, ReactElement } from 'react';
-import classes from './SettingsPage.module.scss';
-
-export const SettingsPage: FC = (): ReactElement => {
-  return <div className={classes.settingsPageBlock}>Settings</div>;
-};
