+++ src/redux/reducers/reducers.ts
@@ -2,8 +2,7 @@ import { combineReducers } from 'redux';
 import type { AnyAction } from 'redux';
 import type { ThunkDispatch } from 'redux-thunk';
 import { profileReducer } from './profile';
-import { messagesReducer } from './messages';
-import { postsReducer } from './posts';
+import { dialogsReducer } from './dialogs';
 import { usersReducer } from './users';
 import { authReducer } from './auth';
 import { loadingReducer } from './loading';
@@ -11,14 +10,13 @@ import { errorReducer } from './error';
 import { chatReducer } from './chat';
 
 export const rootReducer = combineReducers({
-  profile: profileReducer,
-  messages: messagesReducer,
-  posts: postsReducer,
-  users: usersReducer,
   auth: authReducer,
+  profile: profileReducer,
+  dialogs: dialogsReducer,
+  users: usersReducer,
+  chat: chatReducer,
   loading: loadingReducer,
   error: errorReducer,
-  chat: chatReducer,
 });
 
 export type AppState = ReturnType<typeof rootReducer>;
