--- src/tests/reducers/posts.test.ts
@@ -1,36 +1,0 @@
-import { addPost, deletePost } from '../../redux/actions/posts';
-import { postsReducer } from '../../redux/reducers/posts';
-import type { PostsState } from '../../redux/types/posts';
-
-const state: PostsState = {
-  posts: [
-    { id: 1, postText: 'Hi, Max' },
-    { id: 2, postText: 'Hi, Eugene' },
-    { id: 3, postText: 'Hi, Yuri' },
-    { id: 4, postText: 'Hi, Alexei' },
-    { id: 5, postText: 'Hi, Andrey' },
-  ],
-};
-
-describe('Posts reducer tests', (): void => {
-  it('Posts length should be increment', (): void => {
-    const action = addPost('new post text');
-    const newState = postsReducer(state, action);
-
-    expect(newState.posts.length).toBe(6);
-  });
-
-  it('New post text should be correct', (): void => {
-    const action = addPost('new post text');
-    const newState = postsReducer(state, action);
-
-    expect(newState.posts[5].postText).toBe('new post text');
-  });
-
-  it('Posts length should be decrement', (): void => {
-    const action = deletePost(1);
-    const newState = postsReducer(state, action);
-
-    expect(newState.posts.length).toBe(4);
-  });
-});
