+++ src/utils/types/common.ts
@@ -18,10 +18,6 @@ export type ErrorType = {
   code?: number;
   message: string;
 };
-export type DialogType = {
-  id: number;
-  name: string;
-};
 export type MessageType = {
   id: number;
   messageText: string;
